// src/enums/user.ts
export enum ENUM_USER_PERMISSION {
  SUPER_ADMIN = 1,
  ADMIN = 2,
  USER = 3,
  POST_USER = 4,
  PATCH_USER = 5,
  GET_ALL_USER = 6,
  MANAGE_USER = 7,
  MANAGE_LAB_REPORTS = 8,
  MANAGE_PERMISSIONS = 9,
  GET_PERMISSIONS = 11,
  MANAGE_USER_PERMISSIONS = 10,
  GET_USER_PERMISSIONS = 12,
  MANAGE_TEST = 13,
  GET_ALL_TEST = 14
}